+++ app/src/main/java/by/dutov/jee/controllers/servlets/HomeServlet.java
@@ -1,8 +1,7 @@
-package by.dutov.jee.servlets.admin;
+package by.dutov.jee.controllers.servlets;
 
 import by.dutov.jee.utils.CommandServletUtils;
 import lombok.extern.slf4j.Slf4j;
-
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -11,12 +10,12 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @Slf4j
-@WebServlet("/admin/admin")
-public class AdminServlet extends HttpServlet {
+@WebServlet("/home")
+public class HomeServlet extends HttpServlet {
 
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        log.info("Entered admin page.");
-        CommandServletUtils.dispatcher(req, resp, "/admin/adminPage.jsp", true);
+        log.info("Entered Home Page");
+        CommandServletUtils.dispatcher(req, resp, "/homePage.jsp", true);
     }
 }
