+++ app/src/main/java/by/dutov/jee/controllers/servlets/admin/AdminServlet.java
@@ -1,7 +1,8 @@
-package by.dutov.jee.servlets;
+package by.dutov.jee.controllers.servlets.admin;
 
 import by.dutov.jee.utils.CommandServletUtils;
 import lombok.extern.slf4j.Slf4j;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -10,12 +11,12 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @Slf4j
-@WebServlet("/user-info")
-public class UserInfoServlet extends HttpServlet {
+@WebServlet("/admin/admin")
+public class AdminServlet extends HttpServlet {
 
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        log.info("Entered User Info Page");
-        CommandServletUtils.dispatcher(req, resp, "/userInfoPage.jsp", true);
+        log.info("Entered admin page.");
+        CommandServletUtils.dispatcher(req, resp, "/admin/adminPage.jsp", true);
     }
 }
