+++ app/src/main/java/by/dutov/jee/controllers/servlets/student/StudentServlet.java
@@ -1,8 +1,7 @@
-package by.dutov.jee.servlets.teacher;
+package by.dutov.jee.controllers.servlets.student;
 
 import by.dutov.jee.utils.CommandServletUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.slf4j.Slf4j;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,13 +10,13 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/teacher/teacher")
-public class TeacherServlet extends HttpServlet {
-    private static final Logger LOG = LoggerFactory.getLogger(TeacherServlet.class);
+@Slf4j
+@WebServlet("/student/student")
+public class StudentServlet extends HttpServlet {
 
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        LOG.info("Entered Teacher Page");
-        CommandServletUtils.dispatcher(req, resp, "/teacher/teacherPage.jsp", true);
+        log.info("Entered Student Page");
+        CommandServletUtils.dispatcher(req, resp, "/student/studentPage.jsp", true);
     }
 }
