+++ app/src/main/java/by/dutov/jee/repository/person/PersonDAO.java
@@ -1,6 +1,7 @@
 package by.dutov.jee.repository.person;
 
 import by.dutov.jee.people.Person;
+import by.dutov.jee.people.Role;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,5 +12,6 @@ public interface PersonDAO<T> {
     Optional<? extends Person> find(Integer id);
     T update(String name, T t);
     T remove(T t);
-    List<T> findAll();
+    List<? extends Person> findAll(Role role);
+    List<? extends Person> findAll();
 }
