+++ model/src/main/java/by/dutov/jee/people/Teacher.java
@@ -1,6 +1,9 @@
 package by.dutov.jee.people;
 
 import by.dutov.jee.group.Group;
+import com.fasterxml.jackson.annotation.JsonBackReference;
+import com.fasterxml.jackson.annotation.JsonIdentityInfo;
+import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
@@ -11,9 +14,11 @@ import lombok.ToString;
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
+@JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class, property = "id")
 public class Teacher extends Person {
     @ToString.Exclude
     @EqualsAndHashCode.Exclude
+//    @JsonBackReference
     private Group group;
     private double salary;
